+++ src/lib.rs
@@ -36,6 +36,7 @@ mod config;
 mod diagnostics;
 mod eval;
 mod inputs;
+mod test;
 
 /// ignorefile basename to respect.
 const IGNORE_FILENAME: &str = ".sprocketignore";
@@ -136,6 +137,7 @@ async fn inner() -> CommandResult<()> {
         Commands::Validate(args) => commands::validate::validate(args.apply(config)).await,
         Commands::Dev(commands::DevCommands::Doc(args)) => commands::doc::doc(args).await,
         Commands::Dev(commands::DevCommands::Lock(args)) => commands::lock::lock(args).await,
+        Commands::Dev(commands::DevCommands::Test(args)) => commands::test::test(args).await,
     }
 }