+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+* Added an experimental and still under construction `sprocket dev test` command ([#468](https://github.com/stjude-rust-labs/sprocket/pull/468)).
+    * This version merely parses and logs information about found test definitions. **It does not actually execute any WDL code.**
+
 ### Fixed
 
 * `doc` now properly initializes pages in dark mode by default ([#478](https://github.com/stjude-rust-labs/sprocket/pull/478)).